backend/djangorest/src/pyproject.toml
[tool.poetry]
name = "task-app"
version = "0.1.0"
description = ""
authors = ["Someone7777 <85400637+Someone7777@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "task_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.0.8"
djangorestframework = "3.13.1"
django-stubs = "1.12.0"
django-configurations = {version = "2.3.2", extras = ["cache", "database", "email", "search"]}
django-filter = "22.1"
django-cors-headers = "3.13.0"
drf-yasg = "1.21.3"
djangorestframework-simplejwt = "5.2.0"
pyjwt = "2.4.0"
pillow = "9.3.0"
dj-database-url = "1.0.0"
psycopg2 = "2.9.3"
coverage = "6.4.4"
docutils = "0.19"
django-celery-results = "2.4.0"
django-celery-beat = "2.3.0"
redis = "4.3.4"
eventlet = "0.33.1"
uwsgi = "2.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
